requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-import-export==0.2.6
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2